pyproject.toml
[tool.poetry]
name = "Digital Ocean Spaces Export"
version = "0.1.0"
description = "Download all objects from a Digital Ocean 'Spaces' bucket."
authors = ["Todd Birchard"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/toddbirchard/do-spaces-export/"
repository = "https://github.com/toddbirchard/do-spaces-export/"
documentation = "https://github.com/toddbirchard/do-spaces-export/blob/master/README.md"
keywords = [
  "Digital Ocean Spaces",
  "Boto3",
  "Automation",
  "Storage",
  "Scripting"
]


[tool.poetry.dependencies]
python = "^3.9"
boto3 = "*"
python-dotenv = "*"
loguru = "*"

[tool.poetry.dev-dependencies]
black = "*"
isort = "*"
flake8 = "*"
mypy = "*"
pylint = "*"

[tool.poetry.scripts]
run = "main:init_script"

[tool.poetry.urls]
issues = "https://github.com/toddbirchard/do-spaces-export/issues"

[build-system]
requires = ["poetry>=1.1.14"]
build-backend = "poetry.masonry.api"

[tool.pylint.'MESSAGES CONTROL']
disable="C0103,C0301,W0703,W0621"